# app/demo/docs/swagger.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "demo",
    "version": "v0.1.0"
  },
...
  "host": "localhost:5188",
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/contact/add": {
      "post": {
        "summary": "添加好友",
        "operationId": "Contact_Add",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apicontactAddRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apicontactAddReq"
            }
          }
        ],
        "tags": [
          "联系人"
        ]
      }
    },
    "/contact/application/accept": {
      "post": {
        "summary": "同意请求",
        "operationId": "Contact_Accept",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/contactAcceptRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/contactAcceptReq"
            }
          }
        ],
        "tags": [
          "联系人"
        ]
      }
    },
    "/contact/application/list": {
      "post": {
        "summary": "获取好友请求列表",
        "operationId": "Contact_GetApplicationList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/contactGetApplicationListRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/contactGetApplicationListReq"
            }
          }
        ],
        "tags": [
          "联系人"
        ]
      }
    },
    "/contact/application/refuse": {
      "post": {
        "summary": "拒绝请求",
        "operationId": "Contact_Refuse",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/contactRefuseRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/contactRefuseReq"
            }
          }
        ],
        "tags": [
          "联系人"
        ]
      }
    },
    "/contact/application/setRead": {
      "post": {
        "summary": "设置好友请求全部已读",
        "operationId": "Contact_SetApplicationRead",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/contactSetApplicationReadRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/contactSetApplicationReadReq"
            }
          }
        ],
        "tags": [
          "联系人"
        ]
      }
    },
    "/contact/list": {
      "post": {
        "summary": "获取好友列表",
        "operationId": "Contact_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/contactListRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/contactListReq"
            }
          }
        ],
        "tags": [
          "联系人"
        ]
      }
    },
...
  },
  "definitions": {
    "apicontactAddReq": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apicontactAddRsp": {
      "type": "object"
    },
...
    "contactAcceptReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "contactAcceptRsp": {
      "type": "object"
    },
    "contactApplication": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "uid": {
          "type": "string",
          "format": "int64"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "title": "状态[1:审核中 2:通过 3:拒绝]"
        },
        "created_at": {
          "type": "string",
          "format": "int64"
        },
        "updated_at": {
          "type": "string",
          "format": "int64"
        },
        "expires_at": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "contactFriend": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "friend_uid": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "int64"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "alias": {
          "type": "string"
        }
      }
    },
    "contactGetApplicationListReq": {
      "type": "object"
    },
    "contactGetApplicationListRsp": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/contactApplication"
          }
        }
      }
    },
    "contactListReq": {
      "type": "object",
      "properties": {
        "offset": {
          "type": "string",
          "format": "int64"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "contactListRsp": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/contactFriend"
          }
        }
      }
    },
    "contactRefuseReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "contactRefuseRsp": {
      "type": "object"
    },
    "contactSetApplicationReadReq": {
      "type": "object"
    },
    "contactSetApplicationReadRsp": {
      "type": "object"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "swaggerResponseBody": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "detail": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}